requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
python-dotenv
sqlmodel
asyncpg
paho-mqtt
aiomqtt
pydantic-settings
prometheus-client